bit of optimization
