+++ src/software/ai/intent/direct_primitive_intent.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "software/ai/intent/intent.h"
+#include "software/proto/primitive/primitive_msg_factory.h"
 
 /**
  * A DirectPrimitiveIntent is an Intent that directly represents a Primitive, which means
