+++ src/software/proto/primitive/BUILD
@@ -5,7 +5,7 @@ package(default_visibility = ["//visibility:public"])
 cc_library(
     name = "primitive_msg_factory",
     srcs = ["primitive_msg_factory.cpp"],
-    hdrs = ["primitive_msg_factory.h"],
+    hdrs = ["primitive_msg_factory.h","primitive_types.h"],
     deps = [
         "//shared/proto:tbots_cc_proto",
         "//software/geom:angle",
